requirements.txt
fastapi 
uvicorn
python-dotenv
mysql-connector-python
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
bcrypt
python-jose
pydantic